% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_aufbereiten.R
\name{alle_daten_aufbereiten}
\alias{alle_daten_aufbereiten}
\title{Die Daten aufbereiten}
\usage{
alle_daten_aufbereiten(roh)
}
\arguments{
\item{roh}{Die eingelesenen Rohdaten}
}
\value{
Eine Liste mit Werten fuer die Schweiz und Werten fuer alle Anlagen.
* Daten CH `ch`: Alles, was fuer CH gilt, dh alles aus `R_CH` + `Witterung` + BIP etc.
* Daten aller Anlagen `alle`: Alles, was aus dem Sheet `R_Alle`, dh alle Anlagen stammt
* Bezugsjahr `bez`
* Version `vers` Version des Monitoring-Tools
}
\description{
Die eingelesenen Rohdaten aufbereiten. Es gibt eine lange Form mit einer Zeile pro Jahr.
}
