% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_einlesen.R
\name{alle_daten_einlesen}
\alias{alle_daten_einlesen}
\title{Die Daten einlesen}
\usage{
alle_daten_einlesen(pfad)
}
\arguments{
\item{pfad}{Der volle Pfad zum Monitoring-Tool inkl. Dateiname und Endung}
}
\value{
Eine Liste mit
* Daten CH `ch`
* Daten aller Anlagen `alle`
* Daten der Witterung `witt`
* Diverse (BIP, ...?) `div`
* Bezugsjahr `bez`
* Definition der Szenarien `sz`
}
\description{
Wir lesen die Daten aus dem Monitoring-Tool ein
}
