% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{col_vbsa}
\alias{col_vbsa}
\title{Gib eine oder mehrere Farbe(n) aus der VBSA-Palette}
\usage{
col_vbsa(name)
}
\arguments{
\item{name}{Der Name der Farbe oder ein Vector mit mehreren Farben}
}
\value{
Die Farbe(n) als Hex-String, zB '#3E6fA7'
}
\description{
Gib eine oder mehrere Farbe(n) aus der VBSA-Palette
}
