% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_aufbereiten.R
\name{.daten_alle_aufbereiten}
\alias{.daten_alle_aufbereiten}
\title{Daten Alle aufbereiten}
\usage{
.daten_alle_aufbereiten(roh)
}
\arguments{
\item{roh}{Die Rohdaten}
}
\description{
Daten Alle aufbereiten
}
