% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daten_aufbereiten.R
\name{.daten_ch_aufbereiten}
\alias{.daten_ch_aufbereiten}
\title{Daten CH aufbereiten}
\usage{
.daten_ch_aufbereiten(roh)
}
\arguments{
\item{roh}{Die Rohdaten}
}
\description{
Hier werden R_CH, Witterung und BIP zusammengezogen und 'lang' gemacht.
}
