% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zr_helpers.R
\name{.plot_zr_bars}
\alias{.plot_zr_bars}
\title{Bars}
\usage{
.plot_zr_bars(daten, farben, legend_lab = NULL)
}
\arguments{
\item{daten}{Die Daten}

\item{farben}{Die Farben}

\item{legend_lab}{(optional) Die Labels}
}
\description{
Bars
}
