% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tabellen.R
\name{.prep_tabelle}
\alias{.prep_tabelle}
\title{helper für die Tabellen}
\usage{
.prep_tabelle(daten, zeilen, bezugsjahr)
}
\arguments{
\item{daten}{Die Rohdaten Witterung}

\item{zeilen}{Die (korrekten) Namen der Zeilen, die gewählt werden sollen}

\item{bezugsjahr}{Das aktuelle Bezugsjahr bzw. ein Jahr, das in den Daten vorhanden ist.}
}
\description{
helper für die Tabellen
}
