% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{gib_farben_co2_boni}
\alias{gib_farben_co2_boni}
\title{Die Default-Farben fuer den Plot der CO2-Emissionen mit Boni}
\usage{
gib_farben_co2_boni()
}
\description{
Die Default-Farben fuer den Plot der CO2-Emissionen mit Boni
}
