% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gib_formatierten_wert}
\alias{gib_formatierten_wert}
\title{Gib einen formatierten Wert}
\usage{
gib_formatierten_wert(daten, was, jahr, kt = "CH", runder = 0)
}
\arguments{
\item{daten}{Die aufbereiteten Daten, wir erwarten eine Spalte `Was`}

\item{was}{Die Bezeichnung des Wertes (`character`)}

\item{jahr}{Der Wert des Jahres `jahr`}

\item{kt}{(CH) Wenn aus den Daten aller Anlagen gesuch werden soll,
muss hier das Kuerzel der Anlage angegeben werden, zB `AG_1`}

\item{runder}{(optional, default = 0) Die Daten werden auf 0 Kommastellen gerundet.}
}
\description{
Gib einen formatierten Wert
}
