% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gib_spaltennamen}
\alias{gib_spaltennamen}
\title{Gib alle Spaltennamen, dh alle Variablen}
\usage{
gib_spaltennamen(daten)
}
\arguments{
\item{daten}{Die aufbereiteten Daten, `dat$ch` oder `dat$alle`, dh mit einer Spalte `Was`}
}
\description{
Gib alle Spaltennamen, dh alle Variablen
}
