% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farben.R
\name{init_farben}
\alias{init_farben}
\title{Initialisiere die Farben fuer den Bericht}
\usage{
init_farben()
}
\value{
invisible
}
\description{
Es wird die Farb-Palette als 'farb_palette' ins .GlobalEnv geschrieben
}
