% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTplotR-package.r
\docType{package}
\name{MTplotR}
\alias{MTplotR}
\title{MTplotR.}
\description{
Ein Paket, um die Daten aus dem Monitoring-Tool des VBSA
für den jährlichen Bericht einzulesen und die nötigen Grafiken und Tabellen
bereit zu stellen.
}
