% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_co2_boni.R
\name{plot_co2_boni}
\alias{plot_co2_boni}
\title{Title}
\usage{
plot_co2_boni(daten, farben = NULL, y_legend_pos = 0.3)
}
\arguments{
\item{daten}{Die aufbereiteten Daten von `prep_daten_co2_alle_boni`}

\item{farben}{(optional) Ein Vektor mit 6 Farben, default siehe `gib_farben_co2_boni`}

\item{y_legend_pos}{(0.35) Ein Wert zwischen 0 und 1. Er gibt die Höhe der Legende an,
so dass sie exakt auf die 0-Linie ausgerichtet werden kann. Die Wahl des Wertes ist
abhängig von der Grösse der Grafik, dh muss/kann von Hand angepasst werden.}
}
\value{
Ein ggplot2-Objekt
}
\description{
Title
}
