% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grafik.R
\name{plot_grafik}
\alias{plot_grafik}
\title{Eine Grafik plotten}
\usage{
plot_grafik(grafik, hoehe = 2.5, interaktiv = TRUE, testen = FALSE)
}
\arguments{
\item{grafik}{Eine ggplot2-Grafik}

\item{hoehe}{(2.5) Die Höhe der interaktiven Grafik (in inches...), ca. 2/3 von `fig.height` im R-chunk}

\item{interaktiv}{(TRUE) Soll für die `HTML`-Version eine interaktive Grafik sein?}

\item{testen}{(FALSE) Damit wir auch ohne `knitr` bwz. `bookdown` die interaktive Version testen können.}
}
\description{
Eine Grafik plotten
}
