% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hgt.R
\name{plot_hgt_basis}
\alias{plot_hgt_basis}
\title{Der Basis-Plot f체r die HGT-Korrektur}
\usage{
plot_hgt_basis(daten, bezugsjahr, y_max)
}
\arguments{
\item{daten}{Die pr채parierten Daten f체r diesen Plot (mit `prep_daten_hgt`)}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}

\item{y_max}{Der maximale Wert der y-Achse.}
}
\value{
Ein ggplot2-Objekt
}
\description{
Der Basis-Plot f체r die HGT-Korrektur
}
