% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hgt.R
\name{plot_hgt_bezugsjahr}
\alias{plot_hgt_bezugsjahr}
\title{Die W채rmeabgabe im Bezugsjahr mit der HGT-korrigierten Version vergleichen}
\usage{
plot_hgt_bezugsjahr(daten, bezugsjahr, y_max)
}
\arguments{
\item{daten}{Die pr채parierten Daten f체r diesen Plot (mit `prep_daten_hgt`)}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}

\item{y_max}{Das Maximum der y-Achsen. Sinnvollerweise identisch mit `y_max` des HGT-Basisplots}
}
\value{
Ein ggplot2-Objekt
}
\description{
Die W채rmeabgabe im Bezugsjahr mit der HGT-korrigierten Version vergleichen
}
