% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spezifisch.R
\name{plot_spezifisch}
\alias{plot_spezifisch}
\title{Plot spezifische Grössen bzgl. dem verbrannten Abfall}
\usage{
plot_spezifisch(daten)
}
\arguments{
\item{daten}{Eine Liste mit Daten, Labels, Titel, Legendentitel und den Farben}
}
\value{
Ein `ggplot2-`Objekt, es wird _nicht_ geplottet
}
\description{
Plot spezifische Grössen bzgl. dem verbrannten Abfall
}
\examples{
# energie <- list(
#     dat    = filter(dat_spezifisch, Spezifisch \%in\% c("spez_waerme", "spez_strom")),
#     labels = c("Strom", "Wärme"),
#     titel  = "kWh pro Tonne Abfall",
#     fill   = "Spezifische\nEnergieabgabe",
#     farben = col_vbsa(c("orange", "dunkelrot"))
# )
# plot_spezifisch(energie)

}
