% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_temp.R
\name{plot_temp}
\alias{plot_temp}
\title{Die mittelere Temperatur der Schweiz zeigen}
\usage{
plot_temp(daten, y_min, y_max, farben = NULL)
}
\arguments{
\item{daten}{Die aufbereiteten Daten}

\item{y_min}{Minimum der y-Achse}

\item{y_max}{Das Maximum der y-Achse}

\item{farben}{(optional) Zwei Farben: 1. für die Linie (default grün), 2. für den Punkt 2010 (default rot)}
}
\description{
Die mittelere Temperatur der Schweiz zeigen
}
