% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vgl_basisjahr.R
\name{plot_vgl_basisjahr}
\alias{plot_vgl_basisjahr}
\title{Plotte den Vergleich aller Anlagen zwischen Basisjahr und Bezugsjahr}
\usage{
plot_vgl_basisjahr(
  daten,
  farben,
  x_label = NULL,
  y_label = NULL,
  caption_lab = NULL
)
}
\arguments{
\item{daten}{Die präparierten Daten; wir erwarten die drei Spalten `Kurz`, `Jahr`, `Wert`}

\item{farben}{Ein Vektor mit den Farben für die Jahre}

\item{x_label}{optional, Default = `NULL`; die Achsen sind vertauscht, dh x (Anlagen) ist links}

\item{y_label}{optional, Default = `NULL`; die Achsen sind vertauscht, dh y (Werte) ist unten}

\item{caption_lab}{optinal,  Default = `NULL`}
}
\value{
Ein `ggplot2`-Objekt, das geprintet werden kann.
}
\description{
Für jede Anlage werden Balken (horizontal) geplottet.
}
