% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall_alle}
\alias{plot_waterfall_alle}
\title{Der Waterfall- oder Kaskadenplot aller Anlagen der Schweiz}
\usage{
plot_waterfall_alle(
  daten,
  bezugsjahr,
  y_min = NULL,
  y_max = NULL,
  farben = NULL
)
}
\arguments{
\item{daten}{Die pr채parierten Daten mit `prep_daten_waterfall_alle`}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}

\item{y_min}{Der minimale Wert der y-Achse}

\item{y_max}{Der maximale Wert der y-Achse}

\item{farben}{(optional) Die Farben f체r den Plot. Die Anzahl muss exakt mit der Anzahl der S채ulen 체bereinstimmen.}
}
\description{
Der Waterfall- oder Kaskadenplot aller Anlagen der Schweiz
}
