% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waterfall.R
\name{plot_waterfall_ch}
\alias{plot_waterfall_ch}
\title{Plot den Wasserfall CH}
\usage{
plot_waterfall_ch(daten, farben = NULL, shift = 0.003)
}
\arguments{
\item{daten}{Die präparierten Daten. Wir erwarten das Total Vorjahr in der ersten
und das Total Bezugsjahr in der letzten Zeile}

\item{farben}{Die Farben können hier angegeben werden; die Anzahl muss genau stimmen}

\item{shift}{Dieser Wert bestimmt, um wieviel die Grafik um die grossen Werte vertikal zentriert wird.}
}
\value{
Ein `ggplot2`-Objekt, das anschliessend mit `print()` geplottet werden kann.
}
\description{
Plot den Wasserfall CH
}
