% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weitere_emissionsreduktionen.R
\name{plot_weitere_emissionsreduktionen}
\alias{plot_weitere_emissionsreduktionen}
\title{Die weiteren Emissionsreduktionen plotten}
\usage{
plot_weitere_emissionsreduktionen(daten, titel, alpha = 0.6)
}
\arguments{
\item{daten}{Die aufbereiteten Daten}

\item{titel}{Der Titel der Grafik (bzw. die Beschriftung der y-Achse)}

\item{alpha}{Alpha (Transparenz) einstellen, default = 0.5}
}
\value{
Ein `ggplot2-`Objekt, das nicht geplottet wird.
}
\description{
Die weiteren Emissionsreduktionen plotten
}
