% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_zr.R
\name{plot_zr}
\alias{plot_zr}
\title{Plotte eine Zeitreihe}
\usage{
plot_zr(
  daten,
  farben,
  typ = "bars",
  x_label = NULL,
  y_label = NULL,
  y_min = 0,
  y_max = NA,
  legend_lab = NULL,
  caption_lab = NULL
)
}
\arguments{
\item{daten}{Die vorbereiteten Daten.}

\item{farben}{Ein Vector mit den Farben (hex oder Namen)}

\item{typ}{Der Typ der Grafik, default = `bars`. Bisher sind `punkte`, `linien` oder `bars` implementiert.}

\item{x_label}{Beschriftung x-Achse, Default ist leer}

\item{y_label}{Beschriftung y-Achse, Default ist leer}

\item{y_min}{(0) Minimaler Wert der y-Achse, Default = 0}

\item{y_max}{(NA) Maximaler Wert der y-Achse, Default = max(Werte)*1.05}

\item{legend_lab}{(nULL) Beschriftung der Legende, ein Vektor mit den Einträgen. Achtung auf die Reihenfolge! Default ist `NULL`, dann gibt es keine Legende.}

\item{caption_lab}{(NULL) Wir können im Plot unten rechts zB die Herkunft der Daten
(Sheet im Monitoring-Tool) angeben}
}
\value{
Ein ggplot2-Objekt; es wird _nicht_ geplottet.
}
\description{
Wir plotten die vorbereiteten Daten. Achtung: Wenn das Minimum der y-Achse > 0 gewählt wird, werden keine Säulen
geplottet, sondern Punkte.
}
