% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_dat_spezifisch}
\alias{prep_dat_spezifisch}
\title{Die spezifischen Daten aufbereiten}
\usage{
prep_dat_spezifisch(daten)
}
\arguments{
\item{daten}{Die Daten der Schweiz}
}
\value{
Ein `data_frame` für den enstsprechenden Plot
}
\description{
Wir zeigen Wärme/Strom und Fe/NE bezüglich der verbrannten Abfallmenge,
in kWh pro Tonne Abfall bzw. kg Metall pro Tonne Abfall
}
