% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_dat_temp}
\alias{prep_dat_temp}
\title{Die Temperaturdaten Meteo-Schweiz aufbereiten}
\usage{
prep_dat_temp(daten, von)
}
\arguments{
\item{daten}{Die eingelesenen Roh-Daten (ohne Header)}

\item{von}{(Jahr) Ab welchem Jahr soll geplottet werden?}
}
\value{
Ein `data.frame` mit `Jahr`, `Typ`, `Wert`
}
\description{
Sie können anschliessend mit `plot_zr` geplottet werden
}
