% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_dat_weitere_emissionsreduktionen}
\alias{prep_dat_weitere_emissionsreduktionen}
\title{Die Daten der weiteren Emissionsreduktionen aufbereiten}
\usage{
prep_dat_weitere_emissionsreduktionen(daten, max_jahr = 2020)
}
\arguments{
\item{daten}{Die Daten aller Anlagen}

\item{max_jahr}{= 2020,  Bis zu welchem Jahr wollen wir die Säulen zeigen?}
}
\value{
Ein `data_frame` für den entsprechenden Plot
}
\description{
Die totalen Emissionsreduktionen, die im Rahmen von Kompensationsprojekten
geltend gemacht werden, in t CO2-eq. pro Jahr.
}
