% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_daten_co2_alle_boni}
\alias{prep_daten_co2_alle_boni}
\title{Die Daten für den Vergleich der CO2-Emissionen brutto und Boni}
\usage{
prep_daten_co2_alle_boni(daten, bezugsjahr)
}
\arguments{
\item{daten}{Die Daten der Schweiz}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}
}
\value{
Ein `data_frame` wie für die Zeitreihen.
}
\description{
Die Daten für den Vergleich der CO2-Emissionen brutto und Boni
}
