% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_daten_hgt}
\alias{prep_daten_hgt}
\title{Daten für die HGT-Korrektur aufbereiten}
\usage{
prep_daten_hgt(daten, bezugsjahr)
}
\arguments{
\item{daten}{Die Daten der Schweiz (dat$ch)}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}
}
\description{
Daten für die HGT-Korrektur aufbereiten
}
