% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_daten_vgl_vorjahr_alle}
\alias{prep_daten_vgl_vorjahr_alle}
\title{Daten für den Vergleich mit Basisjahr aufbereiten}
\usage{
prep_daten_vgl_vorjahr_alle(daten, spalte, bezugsjahr)
}
\arguments{
\item{daten}{Die aufbereiteten Daten aller Anlagen, dh `dat$alle`}

\item{spalte}{Der Name der Variablen, die verglichen werden soll}

\item{bezugsjahr}{Das aktuelle Bezugsjahr, zB `dat$bez`}
}
\value{
Ein `data_frame` mit den drei Spalten `Kurz`, `Jahr`, `Wert`
}
\description{
Daten für den Vergleich mit Basisjahr aufbereiten
}
