% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_daten_waterfall_alle}
\alias{prep_daten_waterfall_alle}
\title{Daten Wasserfall-Chart aller Anlagen aufbereiten}
\usage{
prep_daten_waterfall_alle(daten, spalte, bezugsjahr)
}
\arguments{
\item{daten}{Die Daten, sie werden hier gefiltert}

\item{spalte}{Der Name der Variablen, die geplottet wird}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}
}
\value{
Ein `data.frame` mit den zwei Spalten `values` und `labels`,
    wie er von `waterfalls` erwartet wird.
}
\description{
Daten Wasserfall-Chart aller Anlagen aufbereiten
}
