% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_daten_waterfall_ch}
\alias{prep_daten_waterfall_ch}
\title{Daten Wasserfall-Chart Schweiz aufbereiten}
\usage{
prep_daten_waterfall_ch(daten, bezugsjahr)
}
\arguments{
\item{daten}{Die (vollständigen) Daten, sie werden hier gefiltert}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}
}
\value{
Ein `data.frame` mit den zwei Spalten `values` und `labels`,
    wie er von `waterfalls` erwartet wird.
}
\description{
Daten Wasserfall-Chart Schweiz aufbereiten
}
