% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_plots.R
\name{prep_daten_zr}
\alias{prep_daten_zr}
\title{Daten für Zeitreihen präparieren}
\usage{
prep_daten_zr(daten, spalten, max_jahr, runder = 2)
}
\arguments{
\item{daten}{Die (vollständigen) Daten}

\item{spalten}{Ein Vector mit den Spaltennamen}

\item{max_jahr}{Bis zu diesem Jahr werden die Daten gezeigt, dh
im Allgemeinen wohl das Bezugsjahr. Es kann aber auch früher oder später sein.}

\item{runder}{(2) Auf wieviele Stellen soll gerundet werden?}
}
\value{
Ein langer `data_frame` mit den Spalten `Typ`, `Jahr`, `Wert` und `TT` (Tooltip)
}
\description{
Daten für Zeitreihen präparieren
}
