% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tabellen.R
\name{prep_tab_waerme_anteil_hgt}
\alias{prep_tab_waerme_anteil_hgt}
\title{Tabelle `waerme_anteil_hgt` vorbereiten}
\usage{
prep_tab_waerme_anteil_hgt(daten)
}
\arguments{
\item{daten}{Die Rohdaten der Witterung}
}
\description{
Diese Daten sind fix aus dem Resultat Rytec
dh sie werden hier von Hand eingegeben und
nicht berechnet
}
