% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_tabellen.R
\name{prep_tab_waerme_korr_zr}
\alias{prep_tab_waerme_korr_zr}
\title{Tabelle `waerme_korr_zr` vorbereiten}
\usage{
prep_tab_waerme_korr_zr(daten, bezugsjahr)
}
\arguments{
\item{daten}{Die Rohdaten der Witterung}

\item{bezugsjahr}{Das aktuelle Bezugsjahr}
}
\description{
Wir erwarten die Rohdaten
}
