% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_zu_breite_tabelle.R
\name{prep_zu_breite_tabelle}
\alias{prep_zu_breite_tabelle}
\title{Eine zu breite Tabelle auf weniger Spalten reduzieren}
\usage{
prep_zu_breite_tabelle(tab, anz_restliche = 3)
}
\arguments{
\item{tab}{Die orignale aufbereitete Tabelle}

\item{anz_restliche}{Wieviele restliche/letzte Spalten sollen gezeigt werden, default = 3}
}
\value{
Die angepasste Tabelle (falls sie effektiv zu breit war)
}
\description{
Eine zu breite Tabelle auf weniger Spalten reduzieren
}
