% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{schreibe_das_bezugsjahr}
\alias{schreibe_das_bezugsjahr}
\title{Schreibe das Bezugsjahr in die .tex-Dateien}
\usage{
schreibe_das_bezugsjahr(jahr, out_dir, file_name = "bezugsjahr.tex")
}
\arguments{
\item{jahr}{Das Bezugsjahr als `integer`}

\item{out_dir}{Der Ordner, wo die .tex-Datei gespeichert wird}

\item{file_name}{Default: `bezugsjahr.tex`, kann hier angepasst werden}
}
\value{
Die Datei wird geschrieben.
}
\description{
Wir schreiben eine tex-Variable mit dem aktuellen Bezugsjahr. Diese wird fuer den Report eingelesen
}
